Clean up test fixture config
