+++ tests/conftest.py
@@ -1,5 +1,6 @@
 import pytest
 
+from fixtures.notification_fixture import smtp_connection
 from fixtures.uaa_fixture import (
     uaa_authenticated,
     uaa_deactivate_user,
