+++ tests/fixtures/notification_fixture.py
@@ -1,0 +1,5 @@
+import pytest
+
+@pytest.fixture(scope="module")
+def smtp_connection():
+    from uaa_bot.notifier import smtplib
